requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==3.4.8
defusedxml==0.7.1
Django==3.0.2
django-admin-thumbnails==0.2.6
django-allauth==0.45.0
easy-thumbnails==2.7.2
idna==3.2
oauthlib==3.1.1
Pillow==8.4.0
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.6
uuid==1.30